pyproject.toml
[project]
name = "scene-tokens"
version = "0.1.0"
requires-python = ">3.10, <3.13"
authors = [
  {name = "Ingrid Navarro", email = "inavarro@stackav.com"},
]
readme = "README.md"

dependencies = [
  "easydict>=1.13",
  "einops>=0.8.1",
  "h5py>=3.13.0",
  "hydra-core>=1.3.2",
  "matplotlib>=3.10.3",
  "numpy>=2.1.3",
  "pandas>=2.2.3",
  "pytorch-lightning>=2.5.1.post0",
  "scikit-learn>=1.6.1",
  "tensorboard>=2.19.0",
  "torch>=2.7.0",
  "tensorflow>=2.19.0",
  "wandb>=0.19.11",
  "tqdm>=4.67.1",
  "joblib>=1.5.1",
  "seaborn>=0.13.2",
  "hydra-colorlog>=1.2.0",
  "omegaconf>=2.3.0",
  "pillow>=11.2.1",
  "pyyaml>=6.0.2",
  "scipy>=1.15.3",
  "google>=3.0.0",
  "typing-extensions>=4.14.0",
  "opencv-python>=4.11.0.86",
  "rich>=14.0.0",
  "pyrootutils>=1.0.4",
  "natsort>=8.4.0",
  "scenario-characterization @ file:///project/scene-tokens/tools/ScenarioCharacterization",
  "mlflow>=3.1.1",
  "pydantic>=2.11.7",
  "pydantic-tensor>=0.2.0",
  "vector-quantize-pytorch>=1.22.18",
  "umap-learn>=0.5.9.post2",
]


[project.optional-dependencies]
dev = [
  "pre-commit>=4.2.0",
]
scenarionet = [
  "metadrive-simulator @ file:///project/scene-tokens/tools/metadrive",
  "scenarionet @ file:///project/scene-tokens/tools/scenarionet",
]

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["tests", "docs", "examples", "scripts"]

[tool.ruff]
exclude = [
    ".eggs",
    ".git",
    ".mypy_cache",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "site-packages",
    "venv",
    # Files ported from other repos, minimally changed.
    "src/scenetokens/utils/data_utils.py",
    "src/scenetokens/models/components",
    "src/scenetokens/datasets/base_dataset.py",
    "src/scenetokens/datasets/scenarionet_dataset.py",
    "src/scenetokens/datasets/scenarionet_types.py",
    "src/scripts/waymo_data_processing.py",
]
line-length = 120
target-version = "py312"
fix = true
lint.ignore = [
  # keep-sorted start
  "COM812",  # flake8-commas - missing-trailing-comma
  "D100",    # pydocstyle - missing docstring in public module
  "D101",    # pydocstyle - missing docstring in public class
  "D102",    # pydocstyle - missing docstring in public method"
  "D103",    # pydocstyle - missing docstring in public function
  "D104",    # pydocstyle - missing docstring in public package
  "D105",    # pydocstyle - missing docstring in magic method
  "D107",    # pydocstyle - missing docstring in init
  "D415",    # pydocstyle - first line should end with a period
  "D203",    # pydocstyle - blank line required before class docstring
  "D205",    # pydocstyle - blank line required before class docstring
  "ERA001",  # eradicate - remove commented-out code
  "FIX002",  # flake8-fixme - line contains todo
  "S101",    # use of assert detected
  "S301",    # pickle can be unsafe to deserialize
  "T20",     # flake8-print
  "TC002",   # type-checking - move third-party import into a type-checking block
  "TD002",   # flake8-todos - missing author in todo
  "TD003",   # flake8-todos - missing issue link on line following todo
]
lint.select = ["ALL"]
lint.fixable = ["ALL"]
lint.unfixable = [
  # Disable the autofixes for the following so that stuff doesn't get deleted when formatting on save
  # keep-sorted start
  "ERA001", # eradicate - remove commented-out code
  "F401",   # pyflakes - unused import
  "F841",   # pyflakes - unused variable
  "RUF100", # unused-noqa
  # keep-sorted end
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = true

[tool.ruff.lint.flake8-type-checking]
exempt-modules = ["typing", "typing_extensions"]
runtime-evaluated-decorators = ["pydantic.dataclasses.dataclass"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.flake8]
max-line-length = 120
